moved TestCompletionProvider into ai
